scramble composable: move stuff to different files
