+++ src/helpers/index.ts
@@ -1,0 +1,3 @@
+export function copyObj(obj: Object) {
+  return JSON.parse(JSON.stringify(obj));
+}
